Corrected wrong man page entry on the -fg option of the tigervncserver.
The -fg option starts the session in the foreground and not the TigerVNC
server. Moreover, this is also the behavior of the vncserver script
shipped with TigerVNC. The documentation of the behavior was simply
wrong in both manpages. Our manpage should now correctly document the
behavior of both the vncserver and the tigervncserver wrapper script
when given the -fg option.